Enhancement PTF
Enhancement to PTF class is returning more data from the API. The superseded by PTF is returned with the PTFR0100 format (which is the basic info format), and just needed to be captured from the output. The superseded PTF chain needed the implementation of the PTFR1000 format. A superseded PTF chain is all the PTFs that are superseded by this PTF. Also updated wording for getSupersedingPTF function to better articulate what it actually returns compared to getSupersededByPTF.

Signed-off-by: Justin Nelson <nelsoncj@us.ibm.com>