Replace all usages of QGY.LIB APIs with their QSYS.LIB equivalents.
In V5R3, the "Open List APIs" were moved from SS1 Option 12 (in library QGY) to SS1 *BASE (in library QSYS). Stub programs were left in QGY to just route the call over to the new QSYS programs, so this has technically continued to work just fine. However, the wording in the V5R3 Memo to Users stated that "These router programs and the library QGY will be removed in a future release. You should change any applications that make library-qualified calls to the Open List APIs in library QGY to call the APIs in library QSYS." As support for connecting to V5R2 was dropped with JTOpen 6.1 (2007), it seems like we can safely do this now. This has the added benefit of removing another dependency on Option 12, and instead shifting that requirement over to the base OS (especially since all these programs did was call things in *BASE, so Option 12 isn't really required for these at all).

Ref: https://public.dhe.ibm.com/systems/power/docs/systemi/v5r3/en_US/rzaq9.pdf (page 52)
Signed-off-by: Parker Young <143426808+pjyoung-ibm@users.noreply.github.com>